added HotSpotProfilingInfo
